Implementation of Serial and start considering delays.
The actual implementation of delays is non-trivial. We do not want to
create calls back and forth without rewinding the stack. This means that
we have to implement two stacks, store the stackpointers and
framepointers and then use setjmp/longjmp to go back and forth.

This can probably all done on the bluenet code side.